pyproject.toml
[tool.poetry]
name = "rag-url-project"
version = "0.1.0"
description = "RAG project with LlamaIndex and OpenAI"
authors = ["Guri"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
llama-index = "^0.10.0"
llama-index-readers-wikipedia = "^0.1.0"
llama-index-program-openai = "^0.1.0"
pydantic = "^2.0.0"
openai = "^1.0.0"
pyyaml = "^6.0.1"
gradio = "<5.0"
huggingface-hub = ">=0.20.0"
streamlit = "^1.28.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"